src-pywebview/pyproject.toml
[project]
name = "hamster-app-py"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "hostage.arith", email = "hostage.ntr@outlook.com"},
]
dependencies = [
    "pywebview>=5.1",
    "cryptography>=43.0.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "Apache-2.0"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.4.2",
    "pyinstaller>=6.9.0",
]
